+++ src/rendering/renderers/FilterRenderer.cpp
@@ -344,8 +344,7 @@ static bool HasComplexPaint(tgfx::Canvas* parentCanvas, const tgfx::Rect& drawin
   auto matrix = parentCanvas->getMatrix();
   matrix.mapRect(&bounds);
   auto surface = parentCanvas->getSurface();
-  auto surfaceBounds = tgfx::Rect::MakeWH(static_cast<float>(surface->width()),
-                                          static_cast<float>(surface->height()));
+  auto surfaceBounds = tgfx::Rect::MakeWH(surface->width(), surface->height());
   bounds.intersect(surfaceBounds);
   auto clip = parentCanvas->getTotalClip();
   if (!clip.contains(bounds)) {
